+++ src/main/kotlin/controller/CarRacingGameController.kt
@@ -2,8 +2,8 @@ package controller
 
 import data.Car
 import data.CarRacingGameDataSource
-import domain.CarGenerator
-import domain.CarRacingGame
+import domain.game.CarRacingGame
+import domain.generator.CarGenerator
 import domain.validator.ValidateUseCases
 import view.InputView
 import view.OutputView
