+++ src/main/kotlin/domain/game/CarRacingGame.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.game
 
 import data.Car
