+++ src/main/kotlin/domain/generator/CarGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.generator
 
 import data.Car
