+++ src/main/kotlin/domain/generator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.generator
 
 interface NumberGenerator {
     fun generate(): Int
