+++ src/main/kotlin/domain/generator/RacingNumberGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.generator
 
 class RacingNumberGenerator : NumberGenerator {
