+++ src/test/kotlin/CarMovingTest.kt
@@ -1,9 +1,9 @@
 import data.Car
 import data.Car.Companion.INITIAL_POSITION
-import domain.CarRacingGame
-import domain.NumberGenerator
-import domain.RacingNumberGenerator.Companion.MINIMUM_NUMBER
-import domain.RacingNumberGenerator.Companion.MINIMUM_NUMBER_TO_MOVE
+import domain.game.CarRacingGame
+import domain.generator.NumberGenerator
+import domain.generator.RacingNumberGenerator.Companion.MINIMUM_NUMBER
+import domain.generator.RacingNumberGenerator.Companion.MINIMUM_NUMBER_TO_MOVE
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
